refactor(interface-cli): extract bootstrap into its own module
Pull the 184-line `bootstrap` function + `Bootstrap` struct +
`build_embedding_provider` helper out of `main.rs` into a new
`crates/interface-cli/src/bootstrap.rs`. This is the shared setup that
builds `StorageLayer`, `SkillRegistry`, `ToolExecutor`, `Orchestrator`,
the LLM provider, and the optional embedding provider for every
subcommand that needs them.

`Bootstrap` struct + all fields flip from private to `pub` so `main()`
can keep destructuring them. The bunch of `use` statements that were
only needed by bootstrap (`OllamaConfig`, `VoyageEmbedder`,
`SkillSource`, `ToolExecutor`, etc.) drop out of `main.rs`.

main.rs shrinks 1965 → 1693 lines (-272). cargo test -p assistant-cli
reports the same 59 lib tests passing; `make lint` is clean.
